requirements.txt
confluent_kafka==1.9.2
numpy==1.24.0
opencv_python==4.5.4.60
python-dotenv==0.21.0
torch==1.13.1
